Separate NavBar into components
